Update README.md
This code is a simple Rock-Paper-Scissors game. The game starts by printing the corresponding ASCII art for rock, paper, and scissors, along with their corresponding number (0, 1, 2) which the player will use to make their choice. The computer will also make a random choice. The game then enters a while loop where the player is prompted to make a choice. If the player inputs 3, the game will exit. If the player inputs a number other than 0, 1, or 2, the game will print that the player has lost. If the player inputs 0, 1, or 2, the game will compare the player's choice to the computer's choice and print the result of the game, "win," "lose," or "draw." If the input is not a number, the game will print "Invalid data!" and continue.